requirements.txt
appdirs==1.4.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.2
cryptography==3.4.7
distlib==0.3.2
Django==3.2
django-cors-headers==3.8.0
django-db-helper==0.0.2
django-filter==2.4.0
django-grpc-framework==0.3.0
django-money==2.0.3
djangogrpcframework==0.2.1
djangorestframework==3.12.4
docutils==0.17.1
drf-yasg==1.20.0
filelock==3.0.12
grpcio==1.41.1
grpcio-tools==1.41.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.0.1
jsonschema==4.2.1
keyring==23.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==20.9
pkginfo==1.7.1
pluggy==1.0.0
protobuf==3.17.2
py==1.11.0
py-moneyed==1.2
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
readme-renderer==29.0
rfc3986==1.5.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.2
tqdm==4.62.2
twine==3.4.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.4.7
webencodings==0.5.1
zipp==3.5.0
